pyproject.toml
[project]
name = "nypd-contract-monitor"
version = "0.1.0"
description = "Monitor NYC contract posting sites for NYPD contracts and send notifications"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "playwright>=1.40",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[tool.setuptools.packages.find]
include = ["contractmonitor*"]

[project.scripts]
contract-monitor = "contractmonitor.cli:main"
